requirements.txt
Django==3.2
djangorestframework==3.14.0
requests==2.28.1
celery
redis
psycopg2-binary
gunicorn
django-celery-beat
python-dotenv
django-redis>=5.4.0
